+++ src/lib/services/types/tx.ts
@@ -12,6 +12,8 @@ import {
   MsgFurtherAction,
   zBechAddr,
   zCoin,
+  zHexAddr20,
+  zHexBig,
   zMessageResponse,
   zPagination,
   zPubkeyMulti,
@@ -375,4 +377,33 @@ export const zTxsCountResponse = z
 export const zBlockTxsResponseSequencer = z.object({
   txs: z.array(zTxsResponseItemFromLcd),
   pagination: zPagination,
+});
+
+// ---------------- JSON RPC ----------------
+export const zEvmTxHashByCosmosTxHashJsonRpc = z.string().transform((val) =>
+  val !== "0x0000000000000000000000000000000000000000000000000000000000000000" // if no related evm tx
+    ? val
+    : null
+);
+
+export const zTxJsonRpc = z.object({
+  blockHash: z.string(),
+  blockNumber: zHexBig,
+  from: z.string(),
+  gas: zHexBig,
+  gasPrice: zHexBig,
+  maxFeePerGas: zHexBig,
+  maxPriorityFeePerGas: zHexBig,
+  hash: z.string(),
+  input: z.string(),
+  nonce: zHexBig,
+  to: zHexAddr20.nullable(),
+  transactionIndex: zHexBig,
+  value: zHexBig,
+  type: z.string(), // TODO: convert to enum later
+  chainId: zHexBig,
+  v: z.string(),
+  r: z.string(),
+  s: z.string(),
+  yParity: z.string(),
 });